Extracted common code
