+++ tests/screenshots/gists.test.js
@@ -1,10 +1,10 @@
-import chai, { expect } from 'chai';
-import sinonChai from 'sinon-chai';
+// import chai, { expect } from 'chai';
+// import sinonChai from 'sinon-chai';
 
-chai.use(sinonChai);
+// chai.use(sinonChai);
 
-describe('Gist Snapshot Testing', () => {
-    it('should be implemented...', () => {
-        expect(true).to.be.true;
-    });
-});
+// describe('Gist Snapshot Testing', () => {
+//     it('should be implemented...', () => {
+//         expect(true).to.be.true;
+//     });
+// });
